added services/added schemas

+++ backend/app/repositories/sources.py
@@ -1,30 +1,38 @@
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy import select
+from sqlalchemy import select, Sequence
 from ..models.sources import Sources
-from typing import List
+from ..schemas.sources import SourcesCreate
 
 
 class SourcesRepository:
     def __init__(self, db:AsyncSession):
         self.db = db
 
-    async def get_all_sources(self) -> List[Sources]:
+    async def get_all_sources(self) -> Sequence[Sources]:
         stmt = select(Sources)
         result = await self.db.execute(stmt)
         return result.scalars().all()
 
-    async def add_source(self, source_data: dict) -> Sources:
+    async def add_source(self, source_schema: SourcesCreate) -> Sources:
+        source_data = source_schema.model_dump()
+
         source = Sources(**source_data)
+
         self.db.add(source)
         await self.db.commit()
         await self.db.refresh(source)
         return source
 
-    async def del_source(self, source_id: int) -> None:
-        stmt = select(Sources).where(source_id==Sources.id)
+    async def del_source(self, source_id: int) -> bool:
+        stmt = select(Sources).where(Sources.id==source_id)
 
         result = await self.db.execute(stmt)
         source = result.scalars().first()
 
+        if not source:
+            return False
+
         await self.db.delete(source)
         await self.db.commit()
+
+        return True
